fix tests

+++ lib/Dialect/AIE/IR/AIEDialect.cpp
@@ -1191,7 +1191,7 @@ LogicalResult BufferOp::verify() {
 
 // FIXME: make address assignment for buffers explicit and move this function to
 // an interface
-uint64_t xilinx::AIE::getBufferBaseAddress(Operation *bufOp) {
+int32_t xilinx::AIE::getBufferBaseAddress(Operation *bufOp) {
   if (auto buf = dyn_cast<BufferOp>(bufOp)) {
     assert(buf.getAddress().has_value() && "buffer must have address assigned");
     return buf.getAddress().value();
